backend/pyproject.toml
[tool.poetry]
name = "story-generator"
version = "0.1.0"
description = "A Python project for generating personalized fairy tales for children using AI."
authors = ["Trinh Quoc Trung <trinhquoctrung@example.com>"]
readme = "README.md"
packages = [{include = "story_generator"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.15"

# Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}

# AI Services
google-generativeai = "^0.8.3"

# Database & Storage
supabase = "^2.10.0"
python-dotenv = "^1.0.1"

# Image Processing
pillow = "^11.0.0"
requests = "^2.32.3"
aiohttp = "^3.11.0"

# Audio
edge-tts = "^7.0.0"
pydub = "^0.25.1"

# Data Validation
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"

# Utilities
python-multipart = "^0.0.20"
huggingface-hub = "^1.1.5"
google-cloud-aiplatform = "^1.128.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^25.1.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"